requirements.txt
asgiref==3.8.1
boto3==1.38.34
botocore==1.38.34
certifi==2025.4.26
charset-normalizer==3.4.2
cloudinary==1.44.0
dj-database-url==2.3.0
Django==5.2.1
django-cloudinary-storage==0.3.0
django-environ==0.12.0
django-storages==1.14.6
django-widget-tweaks==1.5.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
